% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-ncvreg.R
\name{boot.ncvreg}
\alias{boot.ncvreg}
\title{Bootstrap for ncvreg with lasso penalty}
\usage{
boot.ncvreg(
  X,
  y,
  cv_fit,
  lambda,
  sigma2,
  nboot = 100,
  ...,
  cluster,
  seed,
  returnCV = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{X}{The design matrix, without an intercept, as in \code{ncvreg}}

\item{y}{The response vector, as in \code{ncvreg}}

\item{cv_fit}{Instead of \code{X} and \code{y}, an object of type
\code{cv.ncvreg} with \code{returnX = TRUE} and \code{penalty = "lasso"}}

\item{lambda}{A positive scalar value of type numeric, if left unspecified, selected using \code{cv.ncvreg}}

\item{sigma2}{A known / estimated value for the variance used in the Laplace prior and Normal likelihood, if left unspecified, selected using \code{cv.ncvreg} only if \code{lambda} is also left as unspecified}

\item{nboot}{The number of bootstrap iterations. Default is 100.}

\item{...}{}

\item{cluster}{\code{cv.ncvreg} and \code{cv.ncvsurv} can be run in parallel
across a cluster using the \code{parallel} package.  The cluster must be set
up in advance using the \code{makeCluster} function from that package.  The
cluster must then be passed to \code{cv.ncvreg}/\code{cv.ncvsurv} (see
example).}

\item{seed}{You may set the seed of the random number generator in order to
obtain reproducible results. Note that the seed is set at the start of the
code an thus applies to \code{cv.ncvreg} as well as the bootstrap if \code{cv.ncvreg}
is run to select \code{lambda} and estimate \code{sigma2}. If the user would
like to specify a different seed to \code{cv.ncvreg}, they should manually
specify the call to \code{cv.ncvreg} in the arguments.}

\item{returnCV}{Whether to return \code{cv.ncvreg} object, \code{FALSE} by default.}
}
\value{
An object with S3 class \code{boot.ncvreg} containing: \describe {
\item{lowers}{A \code{nboot} by \code{ncol(X)} matrix of the lower bounds
from each bootstrap iteration for each variable.} \item{uppers}{A
\code{nboot} by \code{ncol(X)} matrix of the upper bounds
from each bootstrap iteration for each variable.} \item{modes}{A
\code{nboot} by \code{ncol(X)} matrix of the posterior modes (lasso solution)
from each bootstrap iteration for each variable.} \item{cv.ncvreg}{If
\code{returnCV == TRUE}, an object of type \code{cv.ncvreg} that was fit to
\code{X} and \code{y} to optionally select \code{lambda} and estimate
\code{sigma2}.}}
}
\description{
Perform bootstrapping for the lasso at a single value of the regularization
parameter lambda.
}
\details{
At a specified value of lambda and variance
(which are selected / estimated by default if not provided)
the code does \code{nboot} iterations of taking a pairs bootstrap sample of
\code{X} and \code{y}, fitting \code{ncvreg}, and then uses this fit to
obtain \code{significance_level * 100}\% central credible intervals from
marginal posteriors based on a Laplace prior and a Normal likelihood
specified with partial residuals.
}
\examples{

library(hdrm)
dat <- readData(whoari)

tmp <- boot.ncvreg(dat$X, dat$y)

## Specifying cv_fit
tmp <- boot.ncvreg(cv_fit = cv.ncvreg(dat$X, dat$y, penalty = "lasso", returnX = TRUE))

}
\seealso{
\code{\link{ncvreg}}
}
