% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-ncvreg.R
\name{boot_ncvreg}
\alias{boot_ncvreg}
\title{Bootstrap for ncvreg with lasso penalty}
\usage{
boot_ncvreg(
  X,
  y,
  cv_fit,
  lambda,
  sigma2,
  nboot = 1000,
  ...,
  cluster,
  seed,
  returnCV = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{X}{The design matrix, without an intercept, as in \code{ncvreg}}

\item{y}{The response vector, as in \code{ncvreg}}

\item{cv_fit}{Instead of \code{X} and \code{y}, an object of type
\code{cv.ncvreg} with \code{returnX = TRUE} and \code{penalty = "lasso"}}

\item{lambda}{A positive scalar value of type numeric, if left unspecified,
selected as \code{lambda.min} from \code{cv.ncvreg}}

\item{sigma2}{A known / estimated value for the variance used in obtaining
bootstrap draws, if left unspecified, selected as the \code{cve}
corresponding \code{cv.ncvreg} to \code{lambda}}

\item{nboot}{The number of bootstrap iterations. Default is 1000.}

\item{...}{Additional arguments to \code{ncvreg}/\code{cv.ncvreg}}

\item{cluster}{\code{cv.ncvreg} and the bootstrapping procedure can be run in
parallel across a cluster using the \code{parallel} package. The cluster must
be set up in advance using the \code{makeCluster} function from that package.
The cluster must then be passed to \code{boot_ncvreg} (see example).}

\item{seed}{You may set the seed of the random number generator in order to
obtain reproducible results. Note that the seed is set at the start of the
code an thus applies to \code{cv.ncvreg} as well as the bootstrap if
\code{cv.ncvreg} is run to select \code{lambda} and estimate \code{sigma2}.
If the user would like to specify a different seed to \code{cv.ncvreg}, they
should manually specify the call to \code{cv.ncvreg} in the arguments.}

\item{returnCV}{Whether to return \code{cv.ncvreg} object, \code{FALSE} by
default.}

\item{verbose}{Whether or not to print non-essential messages that highlight
potentially unanticipated behavior.
list(draws = draws, estimates = original_coefs, lambda = lambda, sigma2 = sigma2)}
}
\value{
An object with S3 class \code{boot_ncvreg} containing: \describe {
\item{draws}{A \code{nboot} by \code{ncol(X)} matrix of the bootstrap draws}
\item{estimates}{A length \code{ncol(X)} vector of the estimates from the
lasso model fit on the original data corresponding to \code{lambda}.}
\item{cv.ncvreg}{If \code{returnCV == TRUE}, an object of type
\code{cv.ncvreg} that was fit to \code{X} and \code{y} to select
\code{lambda} and estimate \code{sigma2}. If user supplies \code{cv_fit},
will return the supplied \code{cv.ncvreg} object. If \code{lambda} and
\code{sigma2} are both specified, will return NULL.}}
}
\description{
Perform Hybrid bootstrapping for the lasso at a single value of the
regularization parameter, lambda.
}
\details{
At a specified value of lambda and variance
(which are selected / estimated by default if not provided)
the code does \code{nboot} iterations of taking a pairs bootstrap sample of
\code{X} and \code{y}, fitting \code{ncvreg}, and then uses this fit to
obtain hybrid bootstrap draws.

Note that if a value of lambda is provided, but no value is provided for
sigma2 that sigma2 will be estimated using cross validation. Additionally,
the estimate for sigma2 will be interpolated if the supplied value of lambda
is not in the original sequence of lambda values.
}
\examples{

data(Prostate)

bootfit <- boot_ncvreg(Prostate$X, Prostate$y)

## Specifying cv_fit
bootfit <- boot_ncvreg(cv_fit = cv.ncvreg(Prostate$X, Prostate$y, penalty = "lasso", returnX = TRUE))

## Using a cluster
\dontrun{
library(parallel)
X <- Prostate$X
y <- Prostate$y
cl <- makeCluster(4)
bootfit <- boot_ncvreg(Prostate$X, Prostate$y, cluster = cl)}

}
\seealso{
\code{\link{ncvreg}}
}
