% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-boot-ncvreg.R
\name{ci.boot_ncvreg}
\alias{ci.boot_ncvreg}
\title{Calculating percentile confidence intervals from bootstrap draws}
\usage{
ci.boot_ncvreg(bootfit, alpha = 0.2, quiet = FALSE)
}
\arguments{
\item{bootfit}{An object of type \code{boot_ncvreg}}

\item{alpha}{The desired significance level to obtain confidence intervals
with overall coverage greater than or equal \code{1 - alpha}}

\item{quiet}{If \code{TRUE}, suppress warning that some bootstrap draws are
NA. This occurs when a given variable is \code{singlar} for a given bootstrap
sample (see \code{std}).}
}
\value{
An object with S3 class \code{data.frame} with the columns:
\describe { \item{estimate}{A length \code{ncol(X)} vector of the estimates from the
lasso model fit on the original data corresponding to \code{lambda}.}}
}
\description{
Calculates percentile based confidence intervals from hybrid bootstrap draws
obtained from \code{boot_ncvreg}. This produces intervals that are generally
conservative on average, especially when n is small or the underlying covarites are
especially sparse. However, intervals may undercover when correlation amoung
the predictors is especially high. Although overall coverage is generally
the intervals produced do no attempt to debias and as a result generally
undercover large covariates and overcover covariates near zero.
}
