% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-boot-ncvreg.R
\name{confint.boot_ncvreg}
\alias{confint.boot_ncvreg}
\title{Calculating percentile confidence intervals from Hybrid bootstrap draws}
\usage{
\method{confint}{boot_ncvreg}(boot, alpha = 0.2, quiet = FALSE)
}
\arguments{
\item{boot}{An object of type \code{boot_ncvreg}}

\item{alpha}{The desired significance level to obtain confidence intervals
with overall coverage greater than or equal to \code{1 - alpha}}

\item{quiet}{If \code{TRUE}, suppress warning that some bootstrap draws are
NA. This occurs when a given variable is \code{singlar} for a given bootstrap
sample (see \code{std}).}
}
\value{
An object with S3 class \code{data.frame} with the columns: \describe {
\item{variable}{If the original covariates \code{X} had column names,
these are included as the first column.}
\item{estimate}{A length \code{ncol(X)} vector of the estimates from the
lasso model fit on the original data corresponding to \code{lambda}.}
\item{lower}{A length \code{ncol(X)} vector of the lower bounds from the
Hybrid bootstrap draws with significance level \code{alpha}}
\item{upper}{A length \code{ncol(X)} vector of the upper bounds from the
Hybrid bootstrap draws with significance level \code{alpha}}}
}
\description{
Calculates percentile based confidence intervals from Hybrid bootstrap draws
obtained from \code{boot_ncvreg}. This produces intervals that are generally
conservative on average, especially when n is small or the underlying covarites are
especially sparse. However, intervals may undercover when correlation amoung
the predictors is high. Although overall coverage is generally achieved,
the intervals produced do no attempt to debias and as a result
under cover large covariates and over cover covariates near zero.
}
