\name{fir}
\alias{fir}
\title{Marginal false discovery rates}
\description{Estimates the marginal false discovery rate (mFDR) of a penalized
  regression model.}
\usage{
fir(fit, ...)
}
\arguments{
  \item{fit}{An \code{ncvreg} or \code{ncvsurv} object.}
  \item{...}{Arguments to pass to \code{\link{mfdr}}.}
}
\details{
  This function has been renamed and is currently deprecated.  Use
  \code{\link{mfdr}} instead.
}
