\name{Heart}
\alias{Heart}
\alias{heart}
\title{Risk factors associated with heart disease}
\description{Data from a subset of the Coronary Risk-Factor Study
  baseline survey, carried out in rural South Africa.
  \itemize{
    \item X: A design matrix with 462 observations (rows) and 9
    predictor variables (columns).  The columns are:
    \itemize{
      \item sbp: Systolic blood pressure
      \item tobacco: Cumulative tobacco consumption, in kg
      \item ldl: Low-density lipoprotein cholesterol
      \item adiposity: Adipose tissue concentration
      \item famhist: Family history of heart disease (1=Present, 0=Absent)
      \item typea: Score on test designed to measure type-A behavior
      \item obesity: Obesity
      \item alcohol: Current consumption of alcohol
      \item age: Age of subject
    }
    \item y: Coronary heart disease at baseline; 1=Yes 0=No}}
\usage{data(Heart)}
\source{\url{https://web.stanford.edu/~hastie/ElemStatLearn/}}
\references{
  \itemize{
    \item Hastie T, Tibshirani R, and Friedman J. (2001). \emph{The
      Elements of Statistical Learning}.  Springer.
    \item Rousseauw  J, et al. (1983).  Coronary risk factor screening
    in three rural communities.  \emph{South African Medical Journal},
    \strong{64}, 430-436.
  }
}

\keyword{datasets}
