% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.ncvreg}
\alias{logLik.ncvreg}
\alias{logLik.ncvsurv}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{ncvreg}(object, REML = FALSE, ...)

\method{logLik}{ncvsurv}(object, ...)
}
\arguments{
\item{object}{An \code{ncvreg} or \code{ncvsurv} object, as obtained from \code{\link[=ncvreg]{ncvreg()}} or \code{\link[=ncvsurv]{ncvsurv()}}}

\item{REML}{As in \code{\link[=logLik.lm]{logLik.lm()}}}

\item{...}{For S3 compatibility}
}
\value{
An object with S3 class \code{logLik}. This is a vector of numbers (the log-likelihood at each value
of the regularization parameter) with the following attributes:

\describe{
\item{df}{Degrees of freedom}
\item{nobs}{Number of observations}
}

The 'print' method for 'logLik' objects is not intended to handle vectors; consequently, the
value of the function does not necessarily display correctly. However, it works with 'AIC' and
'BIC' without any glitches and returns the expected vectorized output.
}
\description{
Extract the log-likelihood of an \code{ncvreg} or \code{ncvsurv} object.
}
\details{
Exists mainly for use with \code{\link[stats:AIC]{stats::AIC()}} and \code{\link[stats:AIC]{stats::BIC()}}.
}
\seealso{
\code{\link[stats:logLik]{stats::logLik()}}
}
