% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvfit.R
\name{ncvfit}
\alias{ncvfit}
\title{Direct interface for nonconvex penalized regression (non-pathwise)}
\usage{
ncvfit(
  X,
  y,
  init = rep(0, ncol(X)),
  penalty = c("MCP", "SCAD", "lasso"),
  gamma = switch(penalty, SCAD = 3.7, 3),
  alpha = 1,
  lambda,
  eps = 1e-04,
  max.iter = 1000,
  penalty.factor = rep(1, ncol(X)),
  warn = TRUE
)
}
\description{
This function is intended for users who know exactly what they're doing and want complete control over the fitting process:
no standardization is applied, no intercept is included, no path is fit.
All of these things are best practices for data analysis, so if you are choosing not to do them, you are on your own -- I cannot guarantee that your results will be meaningful.
Some things in particular that you should pay attention to:
\itemize{
\item If your model has an intercept, it is up to you to (un)penalize it properly, typically by settings its corresponding element of \code{penalty.factor} to zero.
\item You should probably provide an initial value; in nonconvex optimization, initial values are very important in determining which local solution an algorithm converges to.
}
}
\details{
To-do: Add family
}
