% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe}
\alias{pipe}
\title{Compute projection-based inference test statistics for all features}
\usage{
pipe(fit, lambda, X = NULL, y = NULL, sigmasq = NULL)
}
\arguments{
\item{fit}{A fitted \code{ncvreg} object.}

\item{lambda}{The value of lambda at which inference should be carried out.}

\item{X, y}{The design matrix and response used to fit the model; in most cases, it is not necessary to provide
\code{X} and \code{y} as they are returned by \code{ncvreg}, but see the \code{returnX} argument in \code{\link[=ncvreg]{ncvreg()}}.}

\item{sigmasq}{The estimated error variance for linear model. Only needed for gaussian case. If not supplied, the residual sums of squares will be computed}
}
\value{
The object returns a data frame with one row per feature and six columns:
\itemize{
\item \code{coef.model}: The coefficient estimate from the penalized regression fit
\item \code{coef.pipe}: The pipe coefficient
\item \code{SE}: Standard error of the pipe coefficient
\item \code{t}: The pipe test statistic, computed based on coef.pipe and SE. It approximately follows a standard normal distribution under the null hypothesis that the
feature is independent of the outcome.
\item \code{p.value}: The pvalue computed based on the test statistic t
\item \code{p.adjust}: The qvalue computed based on Benjamini Hochberg Procedure
}
}
\description{
\code{pipe()} is a function that compute projection-based inference test statistics for all features
}
\examples{
# Linear regression
data(Prostate)
fit <- ncvreg(Prostate$X, Prostate$y)
pipe(fit, 0.1)


# Logistic regression
data(Heart)
X <- Heart$X
y <- Heart$y
fit <- ncvreg(X, y, family='binomial')
pipe(fit, 0.1)
}
