% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boot-ncvreg-r.R
\name{plot.boot.ncvreg.r}
\alias{plot.boot.ncvreg.r}
\title{Title}
\usage{
\method{plot}{boot.ncvreg.r}(eb_boot, n = 30, quiet = TRUE)
}
\arguments{
\item{eb_boot}{}

\item{n}{}

\item{quiet}{}
}
\description{
Title
}
