% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boot-ncvreg.R
\name{plot.boot.ncvreg}
\alias{plot.boot.ncvreg}
\title{Title}
\usage{
\method{plot}{boot.ncvreg}(eb_boot, n = 30, original_order = FALSE, quiet = TRUE)
}
\arguments{
\item{eb_boot}{}

\item{n}{}

\item{quiet}{}
}
\description{
Title
}
