% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-ncvreg.R
\name{boot_ncvreg}
\alias{boot_ncvreg}
\title{Hybrid bootstrap for \code{ncvreg} with Gaussian outcomes}
\usage{
boot_ncvreg(
  X,
  y,
  cv_fit,
  penalty = "lasso",
  lambda,
  gamma = switch(penalty, SCAD = 3.7, 3),
  alpha = 1,
  sigma2,
  nboot = 1000,
  ...,
  cluster,
  seed,
  returnCV = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{X}{The design matrix, without an intercept, as in \code{ncvreg}}

\item{y}{The response vector, as in \code{ncvreg}}

\item{cv_fit}{Instead of \code{X} and \code{y}, an object of type
\code{cv.ncvreg} with \code{returnX = TRUE}}

\item{penalty}{The penalty to be applied to the model.  Either "lasso" (the
default), "MCP", or "SCAD".}

\item{lambda}{A positive scalar value of type numeric, if left unspecified,
selected as \code{lambda.min} from \code{cv.ncvreg}}

\item{gamma}{The tuning parameter of the MCP/SCAD penalty (see details for \code{ncvreg}).
Default is 3 for MCP and 3.7 for SCAD.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the
relative contributions from the lasso/MCP/SCAD penalty and the ridge, or L2
penalty.  \code{alpha=1} is equivalent to lasso/MCP/SCAD penalty, while
\code{alpha=0} would be equivalent to ridge regression.  However,
\code{alpha=0} is not supported; \code{alpha} may be arbitrarily small, but
not exactly 0.}

\item{sigma2}{A known / estimated value for the variance used in obtaining
bootstrap draws, if left unspecified, selected as the \code{cve}
corresponding to \code{lambda} from \code{cv.ncvreg}}

\item{nboot}{The number of bootstrap iterations. Default is 1000.}

\item{...}{Additional arguments to \code{ncvreg}/\code{cv.ncvreg}, must be named}

\item{cluster}{\code{cv.ncvreg} and the bootstrapping procedure can be run in
parallel across a cluster using the \code{parallel} package. The cluster must
be set up in advance using the \code{makeCluster} function from that package.
The cluster must then be passed to \code{boot_ncvreg} (see example).}

\item{seed}{You may set the seed of the random number generator in order to
obtain reproducible results. Note that the seed is set at the start of the
code an thus applies to \code{cv.ncvreg} as well as the bootstrap if
\code{cv.ncvreg} is run to select \code{lambda} and estimate \code{sigma2}.
If the user would like to specify a different seed to \code{cv.ncvreg}, they
should manually specify the call to \code{cv.ncvreg} in the arguments.}

\item{returnCV}{Whether to return \code{cv.ncvreg} object, \code{FALSE} by
default.}

\item{verbose}{Whether or not to print non-essential messages that highlight
potentially unanticipated behavior.}
}
\value{
An object with S3 class \code{boot_ncvreg} containing:
\describe{
\item{draws}{A \code{nboot} by \code{ncol(X)} matrix of the Hybrid bootstrap draws}
\item{estimates}{A length \code{ncol(X)} vector of the estimates from the
lasso model fit on the original data corresponding to \code{lambda}.}
\item{lambda}{The value of the regularization parameter used in obtaining
Hybrid bootstrap draws as selected by \code{cv.ncvreg} or specified by the user.}
\item{sigma2}{The value of the sigma2 used in obtaining
Hybrid bootstrap draws as selected by \code{cv.ncvreg} or specified by the user.}
\item{penalty}{The corresponding penalty that the Hybrid bootstrap draws correspond to.}
\item{alpha}{The tuning parameter used that controls the trade off between the
penalty specified and the Ridge (L2) penalty.}
\item{gamma}{The tuning parameter used for the MCP/SCAD penalty, \code{NULL} if
\code{penalty = "lasso"}}
\item{cv.ncvreg}{If \code{returnCV == TRUE}, an object of type
\code{cv.ncvreg} that was fit to \code{X} and \code{y} to select
\code{lambda} and estimate \code{sigma2}. If user supplies \code{cv_fit},
will return the supplied \code{cv.ncvreg} object. If \code{lambda} and
\code{sigma2} are both specified, will return NULL.}
}
}
\description{
Perform Hybrid bootstrapping for the penalties implemented in \code{ncvreg}
at a single value of the regularization parameter, lambda.
}
\details{
At a specified value of lambda and variance
(which are selected / estimated by default if not provided)
the code does \code{nboot} iterations of taking a pairs bootstrap sample of
\code{X} and \code{y}, fitting \code{ncvreg}, and then uses this fit to
obtain hybrid bootstrap draws.

Note that if a value of lambda is provided, but no value is provided for
sigma2 that sigma2 will be estimated using cross validation. Additionally,
the estimate for sigma2 will be interpolated if the supplied value of lambda
is not in the original sequence of lambda values.
}
\examples{
data(Prostate)

boot <- boot_ncvreg(Prostate$X, Prostate$y)

## Specifying cv_fit
boot <- boot_ncvreg(cv_fit = cv.ncvreg(Prostate$X, Prostate$y, penalty = "lasso", returnX = TRUE))

## Using a cluster
\dontrun{
library(parallel)
X <- Prostate$X
y <- Prostate$y
cl <- makeCluster(4)
boot <- boot_ncvreg(Prostate$X, Prostate$y, cluster = cl)
}
}
\seealso{
\code{\link{ncvreg}}
}
