\name{heart}
\alias{heart}
\title{Risk factors associated with heart disease}
\description{Data from a subset of the Coronary Risk-Factor Study baseline survey, carried out in rural South Africa.  The variables are as follows:
  \itemize{
    \item sbp: Systolic blood pressure
    \item tobacco: Cumulative tobacco consumption, in kg
    \item ldl: Low-density lipoprotein cholesterol
    \item adiposity: Adipose tissue concentration
    \item famhist: Family history of heart disease; either "Present" or "Absent"
    \item typea: Score on test designed to measure type-A behavior
    \item obesity: Obesity
    \item alcohol: Current consumption of alcohol
    \item age: Age of subject
    \item chd: Coronary heart disease at baseline; 1=Yes 0=No}} 
\usage{data(heart)}
\format{A data frame with 462 observations on 10 variables}
\source{http://www-stat-class.stanford.edu/~tibs/ElemStatLearn/}
\references{
  Hastie, T., Tibshirani, R., and Friedman, J. (2001). \emph{The Elements of Statistical Learning}.  Springer.
  
  Rousseauw, J., et al. (1983).  Coronary risk factor screening in three rural communities.  \emph{South African Medical Journal}, \strong{64}, 430-436.}
\keyword{datasets}
