\name{heart}
\alias{heart}
\title{Risk factors associated with heart disease}
\description{The \code{heart} data frame contains 462 rows and 10
  columns.  The data are a subset of the Coronary Risk-Factor Study
  baseline survey, carried out in rural South Africa.}
\usage{data(heart)}
\format{This data frame contains the following columns:
  \describe{
    \item{sbp}{Systolic blood pressure}
    \item{tobacco}{Cumulative tobacco consumption, in kg}
    \item{ldl}{Low-density lipoprotein cholesterol}
    \item{adiposity}{Adipose tissue concentration}
    \item{famhist}{Family history of heart disease; either "Present" or
      "Absent"}
    \item{typea}{Score on test designed to measure type-A behavior}
    \item{obesity}{Obesity}
    \item{alcohol}{Current consumption of alcohol}
    \item{age}{Age of subject}
    \item{chd}{Coronary heart disease at baseline; 1=Yes 0=No}}}
\source{http://www-stat-class.stanford.edu/~tibs/ElemStatLearn/}
\references{
  Hastie, T., Tibshirani, R., and Friedman, J. (2001). \emph{The
    Elements of Statistical Learning}.  Springer.\cr
  Rousseauw, J., et al. (1983).  Coronary risk factor screening in three
  rural communities.  \emph{South African Medical Journal}: 64:
  430-436.\cr}
\keyword{datasets}
