% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.ncvreg}
\alias{logLik.ncvreg}
\alias{logLik.ncvsurv}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{ncvreg}(object, REML = FALSE, ...)

\method{logLik}{ncvsurv}(object, ...)
}
\arguments{
\item{object}{An \code{ncvreg} or \code{ncvsurv} object, as obtained from \code{\link[=ncvreg]{ncvreg()}} or \code{\link[=ncvsurv]{ncvsurv()}}}

\item{REML}{As in \code{\link[=logLik.lm]{logLik.lm()}}}

\item{...}{For S3 compatibility}
}
\description{
Extract the log-likelihood of an \code{ncvreg} or \code{ncvsurv} object.
}
\seealso{
\code{\link[=logLik]{logLik()}}
}
