% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe}
\alias{pipe}
\title{Title}
\usage{
pipe(
  X,
  y,
  fit,
  lambda,
  sigma,
  family = c("gaussian", "binomial", "poisson"),
  penalty = c("MCP", "SCAD", "lasso"),
  gamma = switch(penalty, SCAD = 3.7, 3),
  alpha = 1,
  level = 0.95
)
}
\arguments{
\item{level}{}
}
\description{
Title
}
