% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boot-ncvreg.R
\name{plot.boot_ncvreg}
\alias{plot.boot_ncvreg}
\title{Title}
\usage{
\method{plot}{boot_ncvreg}(
  eb_boot,
  n = 30,
  alpha = 0.2,
  original_order = FALSE,
  absolute_order = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{eb_boot}{}

\item{n}{}

\item{quiet}{}
}
\description{
Title
}
