% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-boot-ncvreg.R
\name{plot.boot_ncvreg}
\alias{plot.boot_ncvreg}
\title{Visualize Hybrid confidence intervals from \code{confint.boot_ncvreg}}
\usage{
\method{plot}{boot_ncvreg}(
  boot,
  n = 30,
  alpha = 0.2,
  order = "original",
  absolute_order = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{boot}{An object of class \code{boot_ncvreg} containing the Hybrid bootstrap samples.}

\item{n}{An integer specifying the number of variables to plot. Default is 30.}

\item{alpha}{A numeric value for the significance level, with default 0.2.}

\item{order}{A character string specifying the order of the variables. Options
are "original", "descending", "ascending", or a vector of variable names.}

\item{absolute_order}{A logical indicating whether to order by the absolute
value of the estimates. Default is FALSE. Only used when order is "descending" or "ascending".}

\item{quiet}{A logical indicating whether to suppress messages during execution of confidence intervals.
Default is TRUE.}
}
\value{
A ggplot object displaying the bootstrap confidence intervals and estimates.
}
\description{
Plots the estimates and Hybrid bootstrap confidence intervals for
an fit from \code{ncvreg} applied to a Gaussian outcome.
}
\examples{
data(Prostate)
boot <- boot_ncvreg(Prostate$X, Prostate$y)
plot(boot, n = 20, alpha = 0.1, order = 'descending', absolute_order = TURE)
}
